6-consumer/lessons/2-language-understanding/code/functions/smart-timer-trigger/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-cognitiveservices-language-luis